w8_1
Changes to path finding still needs a lot of work. also making a fire
route so routes to doors are not cut off. Also will be a criteria for
furniture placement.

Path finding issue on room where a corner is taken out if you start
above the door when you click directly on that doo you will no move
becuase once the path node is equal to the door on the x axis it is a
corner not a floor