w12_3
Start of furniture placement. Right now junk fills spaces.
Need to merge fireroute to new path finding or fix it as there is an
error in room:
Room 1 Door on Right
Room 2 Furthest Right Door
Room 3 First Door on Right
there is junk in middle of hallway blocking door and end of hall.